cleanup

+++ operator/e2e/e2e_test.go
@@ -119,9 +119,11 @@ func TestE2E(t *testing.T) {
 					t.Log(string(out))
 				}
 
-				// cleanup
-				cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/random-yaml-nonroot")
-				require.NoError(t, run(cmd))
+				if !t.Failed() {
+					// cleanup
+					cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/random-yaml-nonroot")
+					require.NoError(t, run(cmd))
+				}
 			},
 		},
 		{
@@ -140,9 +142,11 @@ func TestE2E(t *testing.T) {
 				assert.Equal(t, `"[secret]"`, string(stack.Status.Outputs["secretOutput"].Raw))
 				assert.Equal(t, `"foo"`, string(stack.Status.Outputs["simpleOutput"].Raw))
 
-				// cleanup
-				cmd = exec.Command("bash", "-c", "envsubst < e2e/testdata/git-auth-nonroot/* | kubectl delete -f -")
-				require.NoError(t, run(cmd))
+				if !t.Failed() {
+					// cleanup
+					cmd = exec.Command("bash", "-c", "envsubst < e2e/testdata/git-auth-nonroot/* | kubectl delete -f -")
+					require.NoError(t, run(cmd))
+				}
 			},
 		},
 		{
@@ -158,9 +162,11 @@ func TestE2E(t *testing.T) {
 				assert.Contains(t, stack.Status.Outputs, "targeted")
 				assert.NotContains(t, stack.Status.Outputs, "notTargeted")
 
-				// cleanup
-				cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/targets")
-				require.NoError(t, run(cmd))
+				if !t.Failed() {
+					// cleanup
+					cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/targets")
+					require.NoError(t, run(cmd))
+				}
 			},
 		},
 		{
@@ -185,9 +191,11 @@ func TestE2E(t *testing.T) {
 				_, err = waitFor[autov1alpha1.Workspace]("workspaces/issue-801", "issue-801", 5*time.Minute, "condition=Ready")
 				assert.NoError(t, err)
 
-				// cleanup
-				cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/issue-801")
-				require.NoError(t, run(cmd))
+				if !t.Failed() {
+					// cleanup
+					cmd = exec.Command("kubectl", "delete", "-f", "e2e/testdata/issue-801")
+					require.NoError(t, run(cmd))
+				}
 			},
 		},
 		{
